Move common logic of integration test cases to a common test file
