+++ go-backend/db/organization_test.go
@@ -4,39 +4,12 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	"context"
-	"joshsoftware/peerly/config"
-	"testing"
-	logger "github.com/sirupsen/logrus"
 )
 
 // Define the suite, and absorb the built-in basic suite
 // functionality from testify - including assertion methods.
 type OrganizationTestSuite struct {
 	suite.Suite
-	dbStore Storer
-}
-
-func TestExampleTestSuite(t *testing.T) {
-	suite.Run(t, new(OrganizationTestSuite))
-	suite.Run(t, new(CoreValueTestSuite))
-	suite.Run(t, new(RecognitionHi5TestSuite))
-}
-
-func (suite *OrganizationTestSuite) SetupSuite() {
-	config.Load("application_test")
-
-	err := RunMigrations()
-	if err!=nil {
-		logger.WithField("err", err.Error()).Error("Database migration failed")
-		return
-	}
-
-	store, err := Init()
-	if err != nil {
-		logger.WithField("err", err.Error()).Error("Database init failed")
-		return
-	}
-  suite.dbStore = store
 }
 
 func (suite *OrganizationTestSuite) TestOrganizationsSuccess() {
@@ -53,7 +26,7 @@ func (suite *OrganizationTestSuite) TestOrganizationsSuccess() {
 		Timezone: "IST",
 	}
 	var err error
-	createdOrganization, err := suite.dbStore.CreateOrganization(context.Background(), expectedOrg)
+	createdOrganization, err := dbStore.CreateOrganization(context.Background(), expectedOrg)
 
 	assert.Nil(suite.T(), err)
 
@@ -63,14 +36,14 @@ func (suite *OrganizationTestSuite) TestOrganizationsSuccess() {
 
 	// test list organization
 	var organizationsList []Organization
-	organizationsList, err = suite.dbStore.ListOrganizations(context.Background())
+	organizationsList, err = dbStore.ListOrganizations(context.Background())
 	
 	assert.Nil(suite.T(), err)
 	assert.Equal(suite.T(), organizationsList, []Organization{createdOrganization})
 
 	// test get organization
 	var organizationData Organization
-	organizationData, err = suite.dbStore.GetOrganization(context.Background(), expectedOrg.ID)
+	organizationData, err = dbStore.GetOrganization(context.Background(), expectedOrg.ID)
 	
 	assert.Nil(suite.T(), err)
 	assert.Equal(suite.T(), organizationData, createdOrganization)
@@ -79,16 +52,13 @@ func (suite *OrganizationTestSuite) TestOrganizationsSuccess() {
 	var updatedOrg Organization
 
 	createdOrganization.Name = "Updated name"
-	updatedOrg, err = suite.dbStore.UpdateOrganization(context.Background(), createdOrganization, expectedOrg.ID)
+	updatedOrg, err = dbStore.UpdateOrganization(context.Background(), createdOrganization, expectedOrg.ID)
 	
 	assert.Nil(suite.T(), err)
 	assert.Equal(suite.T(), updatedOrg, createdOrganization)
 
 	//test delete organization
-	err = suite.dbStore.DeleteOrganization(context.Background(), createdOrganization.ID)
+	err = dbStore.DeleteOrganization(context.Background(), createdOrganization.ID)
 	
 	assert.Nil(suite.T(), err)
 }
-
-// TODO run migrations conflicts for multiple test files
-// TODO teardown function
